serial: uartlite: Use simple mechanism to support le and be
commit: effb6752b3 is logically correct but there is a issue in
full relocation of u-boot in microblaze platform. Due to which we are
seeing uartlite getting hung as a result of accessing non-relocated
pointers. So use simple mechanism with the help of a static variable
to use little endian or big endian.

Signed-off-by: Ashok Reddy Soma <ashok.reddy.soma@xilinx.com>
State: pending